src/main/java/br/com/votacao/share/dto/SessaoDto.java
package br.com.votacao.share.dto;

public class SessaoDto {

    private Long sequencial;
    private Long pautaId;
    private String duracao;

    public Long getSequencial() {
        return sequencial;
    }

    public void setSequencial(Long sequencial) {
        this.sequencial = sequencial;
    }

    public Long getPautaId() {
        return pautaId;
    }

    public void setPautaId(Long pautaId) {
        this.pautaId = pautaId;
    }

    public String getDuracao() {
        return duracao;
    }

    public void setDuracao(String duracao) {
        this.duracao = duracao;
    }
}
